requirements.txt
colorama~=0.4.6

pytest~=8.4.1
protobuf~=6.31.1
google-api-python-client~=2.172.0
google-auth-oauthlib~=1.2.2
google-auth-httplib2~=0.2.0